Update colors of images for FRAM
Onboarding images needed to be a different color to get a better contrast with the grey background.
NoFavouriteDeparture needed to be different in light mode vs dark mode. That means that it also has to be added to dark/light folder for AtB and NFK as well, so that the themed component will work correctly.

Partly fixes https://github.com/AtB-AS/kundevendt/issues/3773